+++ .github/workflows/linter.yml
@@ -47,5 +47,5 @@ jobs:
         uses: github/super-linter@v4
         env:
           VALIDATE_ALL_CODEBASE: false
-          DEFAULT_BRANCH: main
+          DEFAULT_BRANCH: feature/login
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
